pyproject.toml
[project]
name = "framemind"
version = "0.1.0"
description = "Production-grade Video Intelligence Engine with CLIP-based frame selection and VLM integration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "FrameMind Team" }]
keywords = ["video", "ai", "vlm", "clip", "computer-vision", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web framework
    "fastapi>=0.109.0,<1.0.0",
    "uvicorn[standard]>=0.27.0,<1.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    
    # Async HTTP client
    "httpx>=0.27.0,<1.0.0",
    
    # Data validation
    "pydantic>=2.6.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    
    # Task queue
    "arq>=0.25.0,<1.0.0",
    
    # Redis
    "redis>=5.0.0,<6.0.0",
    
    # Database
    "aiosqlite>=0.19.0,<1.0.0",
    "sqlalchemy[asyncio]>=2.0.25,<3.0.0",
    
    # ML/CV
    "transformers>=4.37.0,<5.0.0",
    "torch>=2.1.0,<3.0.0",
    "pillow>=10.2.0,<11.0.0",
    "opencv-python-headless>=4.9.0,<5.0.0",
    "numpy>=1.26.0,<2.0.0",
    "scikit-learn>=1.4.0,<2.0.0",
    
    # Video processing
    "ffmpeg-python>=0.2.0,<1.0.0",
    
    # Utilities
    "structlog>=24.1.0,<25.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "aiofiles>=23.2.0,<24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "ruff>=0.2.0,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    "pre-commit>=3.6.0,<4.0.0",
    "httpx>=0.27.0,<1.0.0",
]

s3 = [
    "aioboto3>=12.3.0,<13.0.0",
]

postgres = [
    "asyncpg>=0.29.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
]

[project.scripts]
framemind = "src.api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
